+++ brokerapi/brokers/gcp_service_broker.go
@@ -500,6 +500,18 @@ func getStaticPlans() (map[string][]models.ServicePlan, error) {
 			return map[string][]models.ServicePlan{}, err
 		}
 
+		servicePropertyBytes, err := json.Marshal(p["service_properties"])
+		if err != nil {
+			return map[string][]models.ServicePlan{}, fmt.Errorf("error marshalling service_properties: %s", err)
+		}
+
+		servicePropertyMap := make(map[string]string)
+		if p["service_properties"] != nil {
+			if err := json.Unmarshal(servicePropertyBytes, &servicePropertyMap); err != nil {
+				return map[string][]models.ServicePlan{}, fmt.Errorf("error unmarshalling service_properties: %s, %v", err, servicePropertyBytes)
+			}
+		}
+
 		plan := models.ServicePlan{
 			Name:        planName,
 			Description: p["description"].(string),
@@ -507,12 +519,8 @@ func getStaticPlans() (map[string][]models.ServicePlan, error) {
 				DisplayName: p["display_name"].(string),
 				Bullets:     []string{p["description"].(string), "For pricing information see https://cloud.google.com/pricing/#details"},
 			},
-			ID: id,
-		}
-
-		featureBytes, err := json.Marshal(p["features"])
-		if err != nil {
-			return map[string][]models.ServicePlan{}, fmt.Errorf("error marshalling features: %s", err)
+			ID:                id,
+			ServiceProperties: servicePropertyMap,
 		}
 
 		exists, existingPlan, err := db_service.CheckAndGetPlan(planName, serviceId)
@@ -521,13 +529,13 @@ func getStaticPlans() (map[string][]models.ServicePlan, error) {
 		}
 
 		if exists {
-			existingPlan.Features = string(featureBytes)
+			existingPlan.Features = string(servicePropertyBytes)
 			db_service.DbConnection.Save(&existingPlan)
 		} else {
 			planDetails := models.PlanDetails{
 				ServiceId: serviceId,
 				Name:      p["name"].(string),
-				Features:  string(featureBytes),
+				Features:  string(servicePropertyBytes),
 				ID:        id,
 			}
 			db_service.DbConnection.Create(&planDetails)
@@ -563,10 +571,10 @@ func getDynamicPlans(envVarName string, translatePlanFunc func(details map[strin
 				return []models.ServicePlan{}, "", err
 			}
 
-			// get service-specific plan features from plan interface
-			features := translatePlanFunc(planDetails)
+			// get service-specific plan service_properties from plan interface
+			serviceProperties := translatePlanFunc(planDetails)
 
-			featuresStr, err := json.Marshal(&features)
+			servicePropertiesStr, err := json.Marshal(&serviceProperties)
 			if err != nil {
 				return []models.ServicePlan{}, "", err
 			}
@@ -581,13 +589,13 @@ func getDynamicPlans(envVarName string, translatePlanFunc func(details map[strin
 			// update or make a new plan and save to the database
 			if exists {
 
-				existingPlan.Features = string(featuresStr)
+				existingPlan.Features = string(servicePropertiesStr)
 				db_service.DbConnection.Save(&existingPlan)
 			} else {
 				existingPlan = models.PlanDetails{
 					ServiceId: planDetails["service"],
 					Name:      planDetails["name"],
-					Features:  string(featuresStr),
+					Features:  string(servicePropertiesStr),
 					ID:        id,
 				}
 				db_service.DbConnection.Create(&existingPlan)
@@ -600,7 +608,8 @@ func getDynamicPlans(envVarName string, translatePlanFunc func(details map[strin
 					DisplayName: planDetails["display_name"],
 					Bullets:     []string{planDetails["description"], "For pricing information see https://cloud.google.com/pricing/#details"},
 				},
-				ID: existingPlan.ID,
+				ID:                existingPlan.ID,
+				ServiceProperties: serviceProperties,
 			}
 
 			plansGenerated = append(plansGenerated, plan)