+++ integration_tests/integration_test.go
@@ -205,7 +205,8 @@ var _ = Describe("LiveIntegrationTests", func() {
 		os.Setenv("SERVICES", fakes.Services)
 		os.Setenv("PRECONFIGURED_PLANS", fakes.PreconfiguredPlans)
 
-		os.Setenv("CLOUDSQL_CUSTOM_PLANS", fakes.TestCloudSQLPlan)
+		os.Setenv("CLOUDSQL_MYSQL_CUSTOM_PLANS", fakes.TestCloudSQLMySQLPlan)
+		os.Setenv("CLOUDSQL_POSTGRES_CUSTOM_PLANS", fakes.TestCloudSQLPostgresPlan)
 		os.Setenv("BIGTABLE_CUSTOM_PLANS", fakes.TestBigtablePlan)
 		os.Setenv("SPANNER_CUSTOM_PLANS", fakes.TestSpannerPlan)
 
@@ -229,7 +230,7 @@ var _ = Describe("LiveIntegrationTests", func() {
 	})
 
 	Describe("Broker init", func() {
-		It("should have 9 services in sevices map", func() {
+		It("should have 11 services in sevices map", func() {
 			Expect(len(gcpBroker.ServiceBrokerMap)).To(Equal(9))
 		})
 
@@ -243,7 +244,7 @@ var _ = Describe("LiveIntegrationTests", func() {
 	})
 
 	Describe("getting broker catalog", func() {
-		It("should have 9 services available", func() {
+		It("should have 11 services available", func() {
 			Expect(len(gcpBroker.Services())).To(Equal(9))
 		})
 
@@ -413,6 +414,16 @@ var _ = Describe("LiveIntegrationTests", func() {
 			params := &iamService{
 				serviceId: serviceNameToId[models.StackdriverTraceName],
 				planId:    serviceNameToPlanId[models.StackdriverTraceName],
+			}
+			testIamBasedService(gcpBroker, params)
+		}, timeout)
+	})
+
+	Describe("datastore", func() {
+		It("can provision/bind/unbind/deprovision", func() {
+			params := &iamService{
+				serviceId: serviceNameToId[models.DatastoreName],
+				planId:    serviceNameToPlanId[models.DatastoreName],
 			}
 			testIamBasedService(gcpBroker, params)
 		}, timeout)